requirements.txt
alembic==1.3.3
aniso8601==8.0.0
attrs==19.3.0
backcall==0.1.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
decorator==4.4.1
Flask==1.1.1
Flask-HTTPAuth==3.3.0
Flask-JWT-Extended==3.24.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.1
idna==2.8
importlib-metadata==1.5.0
inflect==4.0.0
ipython==7.11.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.0
Mako==1.1.1
MarkupSafe==1.1.1
more-itertools==8.2.0
packaging==20.1
parso==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.3
ptyprocess==0.6.0
py==1.8.1
Pygments==2.5.2
PyJWT==1.7.1
pyparsing==2.4.6
pytest==5.3.5
python-dateutil==2.8.1
python-dotenv==0.10.5
python-editor==1.0.4
pytz==2019.3
requests==2.22.0
six==1.14.0
SQLAlchemy==1.3.13
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
Werkzeug==0.16.1
zipp==2.1.0
